backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2.3
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
exceptiongroup==1.1.3
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
upgrade-pip==0.1.4
urllib3==2.0.4
